requirements.txt
Flask==2.2.3
gunicorn==21.2.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
